2019 - day 21

+++ 2019/day21/test.h
@@ -3,6 +3,6 @@
 #include <string>
 
 namespace aoc2019_day21 {
-    int part_1(std::string_view path);
-    int part_2(std::string_view path);
+    unsigned long long part_1(std::string_view path);
+    unsigned long long part_2(std::string_view path);
 }
